Use link and move functions to utils
