+++ src/collections/members/ashish-tiwari/index.mdx
@@ -9,7 +9,7 @@ location: New Delhi, India
 bio: I am a final year computer science undergraduate at IP University, Delhi. Passionate Backend Developer with an interest in microservices and cloud-native ecosystems. Constantly learning new things based on interest & curiosity and trying to give back as much in open-source. Other than that, I love solving algorithm problems. A side note about me would be that, once upon a time, I fell in love with mathematics and physics, which still is an ongoing affair.
 badges:
  - meshery
-maintainer: false
+maintainer: no
 blurb: Mesh Head
 status: Active
 published: true
